Add post tuple
Everything should be familiar here except the define-persistent
passage. What "define-persistent" does is just telling Factor which
field belongs to which database column and defining the types of each
column.

define-persistent takes 3 values from the stack: the class, the
database table name and an array of column definitions on top. A
column definition takes the following form:
       { field-name column-name type extra-info }.