Add new-post template
The t:validation-errors will display a list of validation errors
(pretty self-explaining, isn't it? ;)). We will come to validation in
the next commit.

The t:form works nearly the same as a html form tag. The
"$blog/new-post" has the same meaning as before, it's a
dispatcher-relative path.

t:field:
  a html input, the value is saved in "title", the max. length is 100.

t:text-area:
  a html text-area, the value is saved in "content".