Add <new-post-action> word
This action doesn't, unlike the recent-posts action, do anything special
when accessed through a http get request (init field) but has special
quotations for http post requests. The validate and submit quotations
are called when accessed through a http post request (validate first,
then submit).

The validate quotation is quite simple, it just calls validate-post.

The submit quotation has more new stuff. It first creates a new empty
post object and then fills its title and content fields with the values
provided by the html form (remember the t:name="title" in new-post.xml
for example). Then the created-at slot is filled with the current
timestamp. Finally the tuple is written to the db and a redirect
response is pushed onto the stack. The submit quotation always has to
push one response onto the top of the stack.