Add <blog> word
The <blog> word creates a new blog dispatcher. Then it associates the
root path ("/") with the recent-posts action which means everytime the
url "/" is accessed this action will be executed. Next it associates the
"/new-post" url with our new-post action.

Last it wraps the whole dispatcher in a boilerplate object. This
boilerplate is used to factor out the common parts of the templates for
the actions.