Add <view-post-action> word
Again we're using a page-action but this time we use the additional rest
field. The value of the rest field tells the action where to store the
rest of the url which isn't matched by the dispatcher.

In the init quotation we first make sure the value stored in "id" is an
integer. Next we get that value, query for the post object with that
value as its id. from-object does essentially what t:bind-each does,
with the exception that from object just takes one object and not a
sequence.

Last we add this action to our blog dispatcher in <blog>.

This ends the first part of this gitutorial.