Add edit-post template
Everything is the same as the new-post template except for the t:for
attribute of the t:form element. The t:for attribute takes a list of
comma-seperated values which will then be included as hidden fields.

Notice also that we're not sending the post request with the filled form
to the same address as the get request for the form. This is explained
in the next commit.