Add <edit-post-action> and <submit-post-action>
<edit-post-action> does the same as <view-post-action> but uses a
different template so we can just create a new <new-post-action> and
change its template slot.

<submit-post-action> is very similar to <new-post-action> but here its
not possible to reuse <new-post-action> without much extra code so I
just do it the old way. Everything should be familiar here though.

Now for the reason we need to different actions. When we access
/edit-post/1 through a GET request the "id" value is set to 1 because of
the rest slot of <edit-post-action>. But when we access /edit-post
through a POST request (which we would if we wrote
t:action="$blog/edit-post" in the template) Furnace gets confused
because we don't have a rest in the url which can be used to initialize
"id" so "id" is set to f. Instead we have "id" in our POST parameters
which is read earlier and so is overwritten by the f. Therefore we send
to "$blog/submit-post".