Add <delete-post-action> word
This time, we don't use a page-action but an action. page-actions have
an extra template slot which is of no use here because we won't display
anything. Notice we will receive the id by the url and not as an POST
parameter. This is just for simplicity's sake, if we run into problems,
we can still fix them later.