Add authentication
Thanks to furnace.auth.login we have almost no work to add
authentication to the blog. We only need to wrap our dispatcher in the
<blog> word by a login-realm. <login-realm> expects a name and a
dispatcher on the stack, name being on top.

I don't have a SSL cert so I have to switch off https by setting secure
to f.

<login-realm> provides our dispatcher with additional "/login" and
"/logout" paths.